Backend/requirements.txt
absl-py==2.2.2
aiofiles==23.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.7
alembic==1.12.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.39.3
botocore==1.39.3
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dotenv==0.9.9
ecdsa==0.19.1
et-xmlfile==1.1.0
executing==2.1.0
facenet-pytorch==2.6.0
Farama-Notifications==0.0.4
fastapi==0.104.1
fastjsonschema==2.20.0
filelock==3.16.1
firebase-admin==6.9.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==22.0.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
hyperframe==6.1.0
idna==3.10
importlib_metadata==7.1.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
namex==0.0.9
narwhals==1.43.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.6
openai==1.12.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.9.10
overrides==7.7.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.1.0
platformdirs==4.3.6
playwright==1.48.0
plotly==6.2.0
pluggy==1.6.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
PyWavelets==1.7.0
pywin32==307
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.0.7
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.0.0
setuptools==75.1.0
Shimmy==2.0.0
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
stable_baselines3==2.4.1
stack-data==0.6.3
starlette==0.27.0
streamlit==1.46.1
stringzilla==3.12.5
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torch-geometric==2.6.1
torchvision==0.22.1
torchviz==0.0.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typer==0.9.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.168
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.24.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.0.3
wheel==0.45.1
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==3.0.0
xlrd==2.0.1
yarl==1.18.3
zipp==3.19.2
